Split MarkdownChangelogEntry class int subclasses: - MarkdownChangelogEntryMergeRequest - MarkdownChangelogEntryIssues - MarkdownChangelogEntryLinks - MarkdownChangelogEntryAuthor
